final adding

+++ database/migrations/2024_09_14_113032_create_categories_table.php
@@ -14,7 +14,7 @@ return new class extends Migration
         Schema::create('categories', function (Blueprint $table) {
             $table->id();
             $table->string('name');
-            $table->integer('employer_id');
+            $table->unsignedBigInteger('employer_id');
             $table->timestamps();
             //foreign key
             $table->foreign('employer_id')->references('id')->on('users')->onDelete('cascade');
